Refactoring of the code. Removed redundant endpoint :  AddOrgansToDonorAsync Refactoring of Register endpoint
