+++ Donor/Enums/DonorStatus.cs
@@ -1,0 +1,11 @@
+namespace Donor.Enums
+{
+    public enum DonorStatus
+    {
+        Applied = 0,    // Default status
+        Approved = 1,
+        Rejected = 2,
+        OnHold = 3     // Requires a reason
+    }
+
+}
