+++ Donor/Enums/PreferredContactMode.cs
@@ -1,13 +1,12 @@
 
 using System.Text.Json.Serialization;
 
-namespace Donor.Entities
+namespace Donor.Enums
 {
     [JsonConverter(typeof(JsonStringEnumConverter))]
-    public enum Gender
+    public enum PreferredContactMode
     {
-        Male,
-        Female,
-        Other
+        SMS = 0,
+        EMAIL = 1
     }
 }
