+++ Donor/Profiles/OrgansProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using Donor.Entities;
+using Donor.Models;
+
+namespace Donor.Profiles
+{
+	public class OrgansProfile : Profile
+	{
+        public OrgansProfile()
+        {
+			CreateMap<Organ, OrganDto>();
+			CreateMap<OrganDto, Organ>();
+		}
+    }
+}
